refactor: adiciona externalCards em backend
